+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes of GetSimple Legacy CMS will be documented in this file.
 - Display log file as text on the page `/admin/log.php`. Add GET parameter `view=text` to URL.
 - Display content of the PHP errors log file. Use URL `/admin/log.php?log=errorlog.txt&view=text`.
 - Added the field *Website Description* to store the site description on the *Website Settings* page.
+- Added the field *Website Tagline* to store the site slogan or subtitle on the *Website Settings* page.
 - Fill in the *Website URL* field on the *Website Settings* page with the suggested website URL by clicking on it.
 - Control the use of the code editor via *User Settings*. The `GSNOHIGHLIGHT` constant is deprecated.